Move evaluateGuess to separate turn file
