+++ src/card.js
@@ -8,15 +8,7 @@ const createCard = (id, question, answers, correctAnswer) => {
   return card
 }
 
-const evaluateGuess = (guess, correctAnswer) => {
-  if (guess === correctAnswer) {
-    return "correct!"
-  } else {
-    return "incorrect"
-  }
-}
 
 module.exports = {
   createCard,
-  evaluateGuess,
 }
