+++ src/turn.js
@@ -1,0 +1,11 @@
+const evaluateGuess = (guess, correctAnswer) => {
+  if (guess === correctAnswer) {
+    return "correct!"
+  } else {
+    return "incorrect"
+  }
+}
+
+module.exports = {
+  evaluateGuess,
+}
