+++ test/Card-test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const expect = chai.expect;
 
-const { createCard, evaluateGuess } = require('../src/card');
+const { createCard } = require('../src/card');
 
 describe('card', function () {
   it('should be a function', function () {
@@ -18,29 +18,8 @@ describe('card', function () {
   });
 });
 
-describe("guess", () => {
-  it("should be a function", () => {
-    expect(createCard).to.be.a("function");
-  });
-
-  it("should tell user guess is correct", () => {
-    const card = createCard(1, "What allows you to define a set of related information using key-value pairs?", ["object", "array", "function"], "object");
-    const userGuess = "object";
-    let guessResult = evaluateGuess(userGuess, card.correctAnswer);
-    expect(guessResult).to.equal("correct!");
-  });
-
-  it("should tell user guess is incorrect", () => {
-    const card = createCard(1, "What allows you to define a set of related information using key-value pairs?", ["object", "array", "function"], "object");
-    const userGuess = ["array", "function"];
-    let guessResult = evaluateGuess(userGuess, card.correctAnswer);
-    expect(guessResult).to.equal("incorrect");
-  });
-});
-
 module.exports = {
   createCard, 
-  evaluateGuess, 
 }
